refactor: harden security + budget packages (#2257)
