+++ data/feature_index.json
@@ -1199,6 +1199,6 @@
       ]
     }
   ],
-  "generated_at": "2026-06-09T10:08:09.577846Z",
+  "generated_at": "2026-06-08T18:48:17.142880Z",
   "schema_version": 1
 }
