+++ scripts/_module_size_baseline.json
@@ -1,10 +1,10 @@
 {
   "description": "Frozen baseline of oversized src/synthorg/ modules. Each entry under `locations` is `<posix_path>: <current_loc>`. Modules absent from the baseline are enforced strictly against their tier cap. Modules present here may stay at or below the recorded LOC; growing past it fails the gate. Regenerate (rare; requires explicit user approval) with: uv run python scripts/check_module_size_budget.py --update-baseline",
   "locations": {
-    "src/synthorg/api/approval_store.py": 564,
+    "src/synthorg/api/approval_store.py": 561,
     "src/synthorg/api/bus_bridge.py": 528,
     "src/synthorg/api/controllers/activities.py": 598,
-    "src/synthorg/api/controllers/agents.py": 549,
+    "src/synthorg/api/controllers/agents.py": 547,
     "src/synthorg/api/controllers/analytics.py": 626,
     "src/synthorg/api/controllers/custom_rules.py": 514,
     "src/synthorg/api/controllers/departments.py": 709,
@@ -13,21 +13,17 @@
     "src/synthorg/api/controllers/requests.py": 681,
     "src/synthorg/api/controllers/users.py": 604,
     "src/synthorg/api/controllers/workflows.py": 536,
-    "src/synthorg/api/dto.py": 693,
+    "src/synthorg/api/dto.py": 688,
     "src/synthorg/api/lifecycle_helpers/config_apply.py": 542,
     "src/synthorg/api/openapi.py": 574,
-    "src/synthorg/budget/coordination_metrics.py": 580,
-    "src/synthorg/budget/optimizer.py": 679,
-    "src/synthorg/budget/reports.py": 508,
     "src/synthorg/communication/conflict_resolution/human_strategy.py": 521,
     "src/synthorg/communication/meeting/orchestrator.py": 604,
     "src/synthorg/communication/meeting/structured_phases.py": 696,
-    "src/synthorg/config/schema.py": 759,
-    "src/synthorg/core/company.py": 639,
-    "src/synthorg/docs_engine/service.py": 574,
+    "src/synthorg/config/schema.py": 758,
+    "src/synthorg/core/company.py": 637,
     "src/synthorg/hr/activity_service.py": 673,
     "src/synthorg/hr/performance/models.py": 514,
-    "src/synthorg/hr/promotion/service.py": 547,
+    "src/synthorg/hr/promotion/service.py": 544,
     "src/synthorg/hr/scaling/service.py": 522,
     "src/synthorg/hr/training/service.py": 682,
     "src/synthorg/integrations/connections/catalog.py": 685,
@@ -38,7 +34,7 @@
     "src/synthorg/memory/ranking.py": 515,
     "src/synthorg/memory/retriever.py": 662,
     "src/synthorg/memory/self_editing.py": 697,
-    "src/synthorg/memory/tools/knowledge_architect.py": 507,
+    "src/synthorg/memory/tools/knowledge_architect.py": 505,
     "src/synthorg/observability/audit_chain/tsa_client.py": 511,
     "src/synthorg/observability/config.py": 757,
     "src/synthorg/observability/prometheus_collector.py": 684,
@@ -50,14 +46,13 @@
     "src/synthorg/providers/management/dtos.py": 733,
     "src/synthorg/providers/presets.py": 666,
     "src/synthorg/providers/routing/strategies.py": 526,
-    "src/synthorg/security/config.py": 552,
-    "src/synthorg/security/llm_evaluator.py": 639,
-    "src/synthorg/security/safety_classifier.py": 549,
-    "src/synthorg/security/service.py": 561,
     "src/synthorg/settings/dispatcher.py": 575,
     "src/synthorg/settings/resolver.py": 1152,
-    "src/synthorg/templates/presets.py": 558,
     "src/synthorg/templates/renderer.py": 658,
+    "src/synthorg/tools/_git_base.py": 539,
+    "src/synthorg/tools/external_api/external_api_tool.py": 528,
+    "src/synthorg/tools/git_tools.py": 535,
+    "src/synthorg/tools/sandbox/subprocess_sandbox.py": 663,
     "src/synthorg/workers/claim.py": 639,
     "src/synthorg/workers/runtime_builder.py": 836
   }
