+++ src/synthorg/budget/_aggregation.py
@@ -13,6 +13,7 @@ the rejection at a richer scope (with ``agent_id`` / ``task_id`` /
 
 import math
 from collections import defaultdict
+from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
 from synthorg.budget.currency import assert_currencies_match
@@ -20,8 +21,6 @@ from synthorg.constants import BUDGET_ROUNDING_PRECISION
 from synthorg.observability import get_logger
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
-
     from synthorg.budget.cost_record import CostRecord
 
 logger = get_logger(__name__)
