+++ src/synthorg/budget/category_analytics.py
@@ -6,6 +6,7 @@ compute orchestration overhead ratios from cost records tagged with
 """
 
 import math
+from collections.abc import Sequence
 from typing import TYPE_CHECKING, Self
 
 from pydantic import BaseModel, ConfigDict, Field, computed_field, model_validator
@@ -21,8 +22,6 @@ from synthorg.constants import BUDGET_ROUNDING_PRECISION
 from synthorg.observability import get_logger
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
-
     from synthorg.budget.cost_record import CostRecord
 
 logger = get_logger(__name__)
