+++ src/synthorg/security/output_scan_policy.py
@@ -6,21 +6,17 @@ sensitive data -- redact, withhold, log-only, or delegate based on
 autonomy level.
 """
 
+from collections.abc import Mapping
 from types import MappingProxyType
-from typing import TYPE_CHECKING, Protocol, runtime_checkable
+from typing import Protocol, runtime_checkable
 
 from synthorg.core.autonomy_enums import AutonomyLevel
 from synthorg.observability import get_logger
 from synthorg.observability.events.security import (
     SECURITY_OUTPUT_SCAN_POLICY_APPLIED,
 )
-from synthorg.security.models import OutputScanResult, ScanOutcome
-
-if TYPE_CHECKING:
-    from collections.abc import Mapping
-
-    from synthorg.security.autonomy.models import EffectiveAutonomy
-    from synthorg.security.models import SecurityContext
+from synthorg.security.autonomy.models import EffectiveAutonomy
+from synthorg.security.models import OutputScanResult, ScanOutcome, SecurityContext
 
 logger = get_logger(__name__)
