+++ src/synthorg/security/protocol.py
@@ -4,14 +4,13 @@ Defines the async interface that the ``ToolInvoker`` calls for
 pre-tool security checks and post-tool output scanning.
 """
 
-from typing import TYPE_CHECKING, Protocol, runtime_checkable
+from typing import Protocol, runtime_checkable
 
-if TYPE_CHECKING:
-    from synthorg.security.models import (
-        OutputScanResult,
-        SecurityContext,
-        SecurityVerdict,
-    )
+from synthorg.security.models import (
+    OutputScanResult,
+    SecurityContext,
+    SecurityVerdict,
+)
 
 
 @runtime_checkable
