+++ src/synthorg/security/redteam/report_repo.py
@@ -8,16 +8,13 @@ scoped to a single host process. The repo is single-shot per
 """
 
 import asyncio
-from typing import TYPE_CHECKING
 
 from synthorg.core.types import NotBlankStr
 from synthorg.security.redteam.errors import (
     RedTeamReportAlreadyExistsError,
     RedTeamReportNotFoundError,
 )
-
-if TYPE_CHECKING:
-    from synthorg.security.redteam.models import RedTeamReport
+from synthorg.security.redteam.models import RedTeamReport
 
 
 class InMemoryRedTeamReportRepository:
