+++ tests/unit/api/controllers/test_coordination_metrics.py
@@ -4,7 +4,7 @@ from datetime import UTC, datetime, timedelta
 
 import pytest
 
-from synthorg.budget.coordination_metrics import (
+from synthorg.budget.coordination_metric_models import (
     CoordinationMetrics,
     MessageOverhead,
 )
